+++ arviz/plots/backends/matplotlib/posteriorplot.py
@@ -12,6 +12,7 @@ from ...plot_utils import (
     calculate_point_estimate,
     format_sig_figs,
     round_num,
+    vectorized_to_hex,
 )
 from . import backend_kwarg_defaults, backend_show, create_axes_grid, matplotlib_kwarg_dealiaser
 
@@ -35,7 +36,12 @@ def plot_posterior(
     textsize,
     ref_val,
     rope,
+<<<<<<< HEAD
     labeller,
+=======
+    ref_val_color,
+    rope_color,
+>>>>>>> Added rope_color and ref_val_color arguments to plot_posterior
     kwargs,
     backend_kwargs,
     show,
@@ -87,6 +93,8 @@ def plot_posterior(
             skipna=skipna,
             ref_val=ref_val,
             rope=rope,
+            ref_val_color=ref_val_color,
+            rope_color=rope_color,
             ax_labelsize=ax_labelsize,
             xt_labelsize=xt_labelsize,
             **kwargs,
@@ -119,6 +127,8 @@ def _plot_posterior_op(
     skipna,
     ref_val,
     rope,
+    ref_val_color,
+    rope_color,
     ax_labelsize,
     xt_labelsize,
     round_to=None,
@@ -158,15 +168,20 @@ def _plot_posterior_op(
             val,
             format_as_percent(greater_than_ref_probability, 1),
         )
-        ax.axvline(val, ymin=0.05, ymax=0.75, color="C1", lw=linewidth, alpha=0.65)
+        ref_val_props = {
+            "color": vectorized_to_hex("C1")
+            if ref_val_color is None
+            else vectorized_to_hex(ref_val_color)
+        }
+        ax.axvline(val, ymin=0.05, ymax=0.75, lw=linewidth, alpha=0.65, **ref_val_props)
         ax.text(
             values.mean(),
             plot_height * 0.6,
             ref_in_posterior,
             size=ax_labelsize,
-            color="C1",
             weight="semibold",
             horizontalalignment="center",
+            **ref_val_props,
         )
 
     def display_rope():
@@ -191,16 +206,20 @@ def _plot_posterior_op(
             )
         rope_text = [f"{val:.{format_sig_figs(val, round_to)}g}" for val in vals]
 
+        rope_props = {
+            "color": vectorized_to_hex("C2")
+            if rope_color is None
+            else vectorized_to_hex(rope_color)
+        }
         ax.plot(
             vals,
             (plot_height * 0.02, plot_height * 0.02),
             lw=linewidth * 5,
-            color="C2",
             solid_capstyle="butt",
             zorder=0,
             alpha=0.7,
+            **rope_props,
         )
-        text_props = {"size": ax_labelsize, "color": "C2"}
         probability_within_rope = ((values > vals[0]) & (values <= vals[1])).mean()
         ax.text(
             values.mean(),
@@ -208,7 +227,8 @@ def _plot_posterior_op(
             f"{format_as_percent(probability_within_rope, 1)} in ROPE",
             weight="semibold",
             horizontalalignment="center",
-            **text_props,
+            size=ax_labelsize,
+            **rope_props,
         )
         ax.text(
             vals[0],
@@ -216,7 +236,8 @@ def _plot_posterior_op(
             rope_text[0],
             weight="semibold",
             horizontalalignment="right",
-            **text_props,
+            size=ax_labelsize,
+            **rope_props,
         )
         ax.text(
             vals[1],
@@ -224,7 +245,8 @@ def _plot_posterior_op(
             rope_text[1],
             weight="semibold",
             horizontalalignment="left",
-            **text_props,
+            size=ax_labelsize,
+            **rope_props,
         )
 
     def display_point_estimate():